+++ src/main/java/com/yayaveli/inventorymanagement/services/impl/ProviderServiceImpl.java
@@ -1,22 +1,18 @@
 package com.yayaveli.inventorymanagement.services.impl;
 
-import java.util.List;
-import java.util.Optional;
-import java.util.stream.Collectors;
-
 import com.yayaveli.inventorymanagement.dto.ProviderDto;
 import com.yayaveli.inventorymanagement.exceptions.EntityNotFoundException;
 import com.yayaveli.inventorymanagement.exceptions.ErrorCodes;
 import com.yayaveli.inventorymanagement.exceptions.InvalidEntityException;
-import com.yayaveli.inventorymanagement.models.Provider;
 import com.yayaveli.inventorymanagement.repositories.ProviderRepository;
 import com.yayaveli.inventorymanagement.services.ProviderService;
 import com.yayaveli.inventorymanagement.validators.ProviderValidator;
-
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import lombok.extern.slf4j.Slf4j;
+import java.util.List;
+import java.util.stream.Collectors;
 
 @Service
 @Slf4j
@@ -32,8 +28,8 @@ public class ProviderServiceImpl implements ProviderService {
     public ProviderDto save(ProviderDto providerDto) {
         List<String> errors = ProviderValidator.validate(providerDto);
         if (!errors.isEmpty()) {
-            log.error("Provider i not valid {}", providerDto);
-            throw new InvalidEntityException("L'artcle n'est pas valide", ErrorCodes.PROVIDER_NOT_VALID, errors);
+            log.error("Provider is not valid {}", providerDto);
+            throw new InvalidEntityException("Le fournisseur n'est pas valide", ErrorCodes.PROVIDER_NOT_VALID, errors);
         }
         return ProviderDto.fromEntity(providerRepository.save(ProviderDto.toEntity(providerDto)));
     }
@@ -44,12 +40,8 @@ public class ProviderServiceImpl implements ProviderService {
             log.error("Provider is null");
             return null;
         }
-        Optional<Provider> provider = providerRepository.findById(id);
-
-        ProviderDto providerDto = ProviderDto.fromEntity(provider.get());
-
-        return Optional.of(providerDto)
-                .orElseThrow(() -> new EntityNotFoundException("Aucun article avec l'id = " + id,
+        return providerRepository.findById(id).map(ProviderDto::fromEntity)
+                .orElseThrow(() -> new EntityNotFoundException("Aucun fournisseur avec l'id = " + id,
                         ErrorCodes.PROVIDER_NOT_FOUND));
     }
 
@@ -63,6 +55,10 @@ public class ProviderServiceImpl implements ProviderService {
         if (id == null) {
             log.error("Provider is null");
             return;
+        }
+        if(!providerRepository.existsById(id)) {
+            throw new EntityNotFoundException("Aucun fournisseur avec l'id = " + id,
+                    ErrorCodes.PROVIDER_NOT_FOUND);
         }
         providerRepository.deleteById(id);
     }
